IntegratedDriver/app/legacy_ld/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# GUI.py: 
PySimpleGUI == 4.15.1
matplotlib == 3.0.3
pyserial == 3.4
